Correction of Currency, SubCurrency and Session data model

+++ RPGAapp/Data Classes/SubCurrency+CoreDataProperties.swift
@@ -17,6 +17,6 @@ extension SubCurrency {
     }
 
     @NSManaged public var name: String?
-    @NSManaged public var rate: Double
+    @NSManaged public var rate: Int16
 
 }
